requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
slack-bolt>=1.18.0
slack-sdk>=3.23.0
anthropic>=0.7.0
schedule>=1.2.0
selenium>=4.0.0
webdriver-manager>=3.8.0
